# docs/swagger.yaml
basePath: /
definitions:
...
  internal_api.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_api.tokenError:
    properties:
      error:
        type: string
    type: object
  internal_api.tokenRequest:
    properties:
      secret:
        description: 'Secret is the value printed at server startup ("API secret:
          …").'
        type: string
    required:
    - secret
    type: object
  internal_api.tokenResponse:
    properties:
      token:
        type: string
    type: object
info:
  contact: {}
  description: |-
    REST API for managing listeners, reverse-shell sessions, file transfer, and tunnels.
    Every endpoint except /api/v1/auth/token requires a Bearer token obtained via that endpoint.
  title: Platypus API
  version: "1.0"
paths:
...
  /api/v1/auth/token:
    post:
      consumes:
      - application/json
      description: 'Bootstraps a session. The returned token must be sent as `Authorization:
        Bearer <token>` on every other endpoint.'
      parameters:
      - description: Server secret
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.tokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.tokenError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.tokenError'
      summary: Exchange secret for token
      tags:
      - auth
...
  /api/v1/sessions/{id}/files:
    get:
      description: Stream a slice of a remote file as application/octet-stream. Omit
        size=0 to read the whole file; otherwise read [offset, offset+size).
      parameters:
      - description: Session hash
        in: path
        name: id
        required: true
        type: string
      - description: Absolute path to the file
        in: query
        name: path
        required: true
        type: string
      - default: 0
        description: Byte offset to start reading from
        in: query
        name: offset
        type: integer
      - default: 0
        description: Number of bytes to read; 0 = whole file
        in: query
        name: size
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
      security:
      - BearerAuth: []
      summary: Read file
      tags:
      - files
    post:
      consumes:
      - application/octet-stream
      description: Upload bytes to a remote path. Use append=true to chunk large uploads.
        Only Termite sessions support file upload.
      parameters:
      - description: Session hash (must be a Termite session)
        in: path
        name: id
        required: true
        type: string
      - description: Absolute path to write to
        in: query
        name: path
        required: true
        type: string
      - default: false
        description: If true, appends to the file instead of truncating
        in: query
        name: append
        type: boolean
      - description: Raw file bytes
        in: body
        name: body
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: status + bytes_written
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
      security:
      - BearerAuth: []
      summary: Write file
      tags:
      - files
  /api/v1/sessions/{id}/files/size:
    get:
      description: Stat a remote file and return its size in bytes. Requires the session
        to be connected.
      parameters:
      - description: Session hash
        in: path
        name: id
        required: true
        type: string
      - description: Absolute path to the file
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: status + size
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
      security:
      - BearerAuth: []
      summary: File size
      tags:
      - files
...
securityDefinitions:
  BearerAuth:
    description: Value should be "Bearer <token>". Fetch a token via POST /api/v1/auth/token
      using the secret printed at server startup.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"